lint and refactor
