+++ tests/integration/test_charm_with_proxy.py
@@ -6,6 +6,12 @@ import logging
 import subprocess
 from asyncio import sleep
 from pathlib import Path
+from test.integration.helpers.common import (
+    ensure_charm_has_runner,
+    get_runner_names,
+    reconcile,
+    run_in_lxd_instance,
+)
 from typing import AsyncIterator, Optional
 from urllib.parse import urlparse
 
@@ -23,12 +29,6 @@ from charm_state import (
     TOKEN_CONFIG_NAME,
     USE_APROXY_CONFIG_NAME,
     VIRTUAL_MACHINES_CONFIG_NAME,
-)
-from tests.integration.helpers import (
-    ensure_charm_has_runner,
-    get_runner_names,
-    reconcile,
-    run_in_lxd_instance,
 )
 from tests.status_name import ACTIVE
 from utilities import execute_command
@@ -94,20 +94,20 @@ async def app_with_prepared_machine_fixture(
     proxy: str,
 ) -> Application:
     """Application with proxy setup and firewall to block all other network access."""
-    await model.set_config(
-        {
-            "apt-http-proxy": proxy,
-            "apt-https-proxy": proxy,
-            "apt-no-proxy": NO_PROXY,
-            "juju-http-proxy": proxy,
-            "juju-https-proxy": proxy,
-            "juju-no-proxy": NO_PROXY,
-            "snap-http-proxy": proxy,
-            "snap-https-proxy": proxy,
-            "snap-no-proxy": NO_PROXY,
-            "logging-config": "<root>=INFO;unit=DEBUG",
-        }
-    )
+    # await model.set_config(
+    #     {
+    #         "apt-http-proxy": proxy,
+    #         "apt-https-proxy": proxy,
+    #         "apt-no-proxy": NO_PROXY,
+    #         "juju-http-proxy": proxy,
+    #         "juju-https-proxy": proxy,
+    #         "juju-no-proxy": NO_PROXY,
+    #         "snap-http-proxy": proxy,
+    #         "snap-https-proxy": proxy,
+    #         "snap-no-proxy": NO_PROXY,
+    #         "logging-config": "<root>=INFO;unit=DEBUG",
+    #     }
+    # )
 
     machine = await model.add_machine(constraints={"root-disk": 15}, series="jammy")
     # Wait until juju agent has the hostname of the machine.
