+++ tests/integration/test_debug_ssh.py
@@ -4,6 +4,7 @@
 """Integration tests for github-runner charm with ssh-debug integration."""
 import logging
 from datetime import datetime, timedelta
+from test.integration.helpers.common import dispatch_workflow, get_job_logs, get_workflow_runs
 
 from dateutil.tz import tzutc
 from github.Branch import Branch
@@ -13,7 +14,6 @@ from juju.application import Application
 from juju.model import Model
 
 from charm_state import DENYLIST_CONFIG_NAME
-from tests.integration.helpers import dispatch_workflow, get_job_logs, get_workflow_runs
 from tests.status_name import ACTIVE
 
 logger = logging.getLogger(__name__)
