+++ tests/integration/test_openstack.py
@@ -3,6 +3,11 @@
 
 """Integration tests for OpenStack integration."""
 
+from test.integration.helpers.common import (
+    DISPATCH_E2E_TEST_RUN_WORKFLOW_FILENAME,
+    dispatch_workflow,
+)
+
 import openstack.connection
 import pytest
 from github.Branch import Branch
@@ -11,8 +16,6 @@ from github.WorkflowRun import WorkflowRun
 from juju.application import Application
 from juju.model import Model
 from openstack.compute.v2.server import Server
-
-from tests.integration.helpers import DISPATCH_E2E_TEST_RUN_WORKFLOW_FILENAME, dispatch_workflow
 
 
 # 2024/03/19 - The firewall configuration on openstack will be implemented by follow up PR on
